+++ tests/int_bitbucket.rs
@@ -112,6 +112,8 @@ fn test_bitbucket_remote_scan() -> Result<()> {
             max_results: 100,
             slack_query: None,
             slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+            teams_query: None,
+            teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
             s3_bucket: None,
             s3_prefix: None,
             role_arn: None,